pi-daemon/requirements.txt
python-dotenv
opencv-python
